Constructed methyl_surro Object with Complete Data

A methyl_surro object constructed from beta_matrix_comp and wts_vec_lin.

Format

An object of class methyl_surro, containing the following components:

methyl

A matrix of methylation data with CpG sites as row names. Some values are missing, simulating incomplete data scenarios.

weights

A named numeric vector representing the weights for the CpG sites, excluding the intercept.

intercept

The numeric value of the intercept term, or NULL if not applicable.

Examples

data(methyl_surro_comp)
str(methyl_surro_comp)
#> List of 3
#>  $ methyl   : num [1:10, 1:5] 0.288 0.9 0.89 0.963 0.143 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:10] "cg02" "cg07" "cg08" "cg13" ...
#>   .. ..$ : chr [1:5] "samp1" "samp2" "samp3" "samp4" ...
#>  $ weights  : Named num [1:10] -0.00908 -0.00116 0.00598 -0.00756 0.00122 ...
#>   ..- attr(*, "names")= chr [1:10] "cg02" "cg03" "cg06" "cg07" ...
#>  $ intercept: Named num 1.21
#>   ..- attr(*, "names")= chr "Intercept"
#>  - attr(*, "class")= chr "methyl_surro"